refactor: extract fmt module
